+++ baybe/parameters/categorical.py
@@ -19,7 +19,7 @@ class CategoricalParameter(DiscreteParameter):
     """Parameter class for categorical parameters."""
 
     # class variables
-    is_numeric: ClassVar[bool] = False
+    is_numerical: ClassVar[bool] = False
     # See base class.
 
     # object variables