+++ baybe/parameters/substance.py
@@ -35,7 +35,7 @@ class SubstanceParameter(DiscreteParameter):
     """
 
     # class variables
-    is_numeric: ClassVar[bool] = False
+    is_numerical: ClassVar[bool] = False
     # See base class.
 
     # object variables