+++ CHANGELOG.md
@@ -44,6 +44,7 @@ _ `_optional` subpackage for managing optional dependencies
   `SubspaceContinuous.sample_uniform`
 - `SubspaceContinuous.samples_full_factorial` has been replaced with
   `SubspaceContinuous.sample_from_full_factorial`
+- `Parameter.is_numeric` has been replaced with `Parameter.is_numerical`
 
 ## [0.9.1] - 2024-06-04
 ### Changed